src_repo/requirements.txt
tensorflow-gpu==1.13.2
pandas
matplotlib
pillow
seaborn
tensorboard